+++ current/dox/convert_xy.c
@@ -3,32 +3,33 @@
 #include <stdio.h>
 #include <math.h>
 
-int main(void)
+int32_t main(void)
 {
-urg_t urg;
-long *length_data = NULL;
-int length_data_size;
-int i;
-// \~japanese ‹——£ƒf[ƒ^‚ğ X-Y À•WŒn‚É•ÏŠ·‚µ‚Ä•\¦‚·‚é
-// \~english Converts data to X-Y coordinates and displays it
+    urg_t urg;
+    int32_t *length_data = NULL;
+    int32_t length_data_size;
+    int32_t i;
+    // \~japanese è·é›¢ãƒ‡ãƒ¼ã‚¿ã‚’ X-Y åº§æ¨™ç³»ã«å¤‰æ›ã—ã¦è¡¨ç¤ºã™ã‚‹
+    // \~english Converts data to X-Y coordinates and displays it
 
-length_data_size = urg_get_distance(&urg, length_data, NULL);
-for (i = 0; i < length_data_size; ++i) {
-    // \~japanese ‚»‚Ì‹——£ƒf[ƒ^‚Ìƒ‰ƒWƒAƒ“Šp“x‚ğ‹‚ßAX, Y ‚ÌÀ•W’l‚ğŒvZ‚·‚é
-    // \~english Gets the angle in radians for range data, and convert to X-Y coordinates
-    double radian;
-    long length;
-    long x;
-    long y;
+    length_data_size = urg_get_distance(&urg, length_data, NULL);
+    for (i = 0; i < length_data_size; ++i)
+    {
+        // \~japanese ãã®è·é›¢ãƒ‡ãƒ¼ã‚¿ã®ãƒ©ã‚¸ã‚¢ãƒ³è§’åº¦ã‚’æ±‚ã‚ã€X, Y ã®åº§æ¨™å€¤ã‚’è¨ˆç®—ã™ã‚‹
+        // \~english Gets the angle in radians for range data, and convert to X-Y coordinates
+        double radian;
+        int32_t length;
+        int32_t x;
+        int32_t y;
 
-    radian = urg_index2rad(&urg, i);
-    length = length_data[i];
-    // \todo check length is valid
+        radian = urg_index2rad(&urg, i);
+        length = length_data[i];
+        // \todo check length is valid
 
-    x = (long)(length * cos(radian));
-    y = (long)(length * sin(radian));
-    printf("(%ld, %ld), ", x, y);
-}
-printf("\n");
-return 0;
+        x = (int32_t)(length * cos(radian));
+        y = (int32_t)(length * sin(radian));
+        printf("(%ld, %ld), ", x, y);
+    }
+    printf("\n");
+    return 0;
 }
