+++ current/samples/cpp/get_distance_io.cpp
@@ -12,42 +12,40 @@
 #include "Connection_information.h"
 #include "math_utilities.h"
 #include <iostream>
-#include <sstream> 
+#include <sstream>
 #include <bitset>
 
 using namespace qrk;
 using namespace std;
 
-
 namespace
 {
-    void print_data(const Urg_driver& urg,
-                    const vector<long>& data, const vector<long>& io, long time_stamp)
+    void print_data(const Urg_driver &urg,
+                    const vector<int32_t> &data, const vector<int32_t> &io, int32_t time_stamp)
     {
-		cout << "timestamp: " << time_stamp << " [msec]" << endl;
+        cout << "timestamp: " << time_stamp << " [msec]" << endl;
 
-	// \~japanese IO情報を表示
-	// \~english Display IO information
-		stringstream ss;
-		ss << bitset<18>(io[0]);
-		string input = ss.str();
-		stringstream ss2;
-		ss2 << bitset<18>(io[1]);
-		string output = ss2.str();
-		cout << "input    : " << input  << " (" << io[0] << ")" << endl;
-		cout << "output   : " << output << " (" << io[1] << ")" << endl;
+        // \~japanese IO情報を表示
+        // \~english Display IO information
+        stringstream ss;
+        ss << bitset<18>(io[0]);
+        string input = ss.str();
+        stringstream ss2;
+        ss2 << bitset<18>(io[1]);
+        string output = ss2.str();
+        cout << "input    : " << input << " (" << io[0] << ")" << endl;
+        cout << "output   : " << output << " (" << io[1] << ")" << endl;
 
-    // \~japanese 前方のデータのみを表示
-    // \~english Shows only the front step
-        int front_index = urg.step2index(0);
+        // \~japanese 前方のデータのみを表示
+        // \~english Shows only the front step
+        int32_t front_index = urg.step2index(0);
         cout << "distance : " << data[front_index] << " [mm]" << endl;
-		
+
         cout << endl;
     }
 }
 
-
-int main(int argc, char *argv[])
+int32_t main(int32_t argc, char *argv[])
 {
     Connection_information information(argc, argv);
 
@@ -56,7 +54,8 @@ int main(int argc, char *argv[])
     Urg_driver urg;
     if (!urg.open(information.device_or_ip_name(),
                   information.baudrate_or_port_number(),
-                  information.connection_type())) {
+                  information.connection_type()))
+    {
         cout << "Urg_driver::open(): "
              << information.device_or_ip_name() << ": " << urg.what() << endl;
         return 1;
@@ -69,14 +68,19 @@ int main(int argc, char *argv[])
     // \~english Case where the measurement range (start/end steps) is defined
     urg.set_scanning_parameter(urg.deg2step(-90), urg.deg2step(+90), 0);
 #endif
-    enum { Capture_times = 10 };
+    enum
+    {
+        Capture_times = 10
+    };
     urg.start_measurement(Urg_driver::Distance_io, Urg_driver::Infinity_times, 0);
-    for (int i = 0; i < Capture_times; ++i) {
-        vector<long> data;
-		vector<long> io;
-        long time_stamp = 0;
+    for (int32_t i = 0; i < Capture_times; ++i)
+    {
+        vector<int32_t> data;
+        vector<int32_t> io;
+        int32_t time_stamp = 0;
 
-        if (!urg.get_distance_io(data, io, &time_stamp)) {
+        if (!urg.get_distance_io(data, io, &time_stamp))
+        {
             cout << "Urg_driver::get_distance_io(): " << urg.what() << endl;
             return 1;
         }
